大三上/虚拟现实技术/虚拟现实技术2021秋-小作业1/code/requirements.txt
numpy
Pillow
PyOpenGL
glfw
